Add harvest routine
The harvest routine does the following. From the list of repositories it
fetches the CKAN URL entries and then iterates each repository. In the
iteration all the datasets are queried with single requests to get get
all the details. From the JSON result a dataset object is created and
indexed by Elasticsearch.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>