Add logger and extended JSON parser
For better standard output management I have added the Ruby logger to
the harvester. Also there is a fundamental problem with the delivered
metadata. Some properties are actually arrays or objects, but are
encoded as strings.

To overcome this difficulty I have prototyped a method which recursively
applies JSON.parse.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>