Make harvester threaded
For faster harvesting I am now using Sidekiq to parallelize the tasks.
This adds the overhead of using Redis and a Sidekiq server. The result,
however, is quite fast.

At the moment for each HTTP request a worker is spawned which generates
to much traffic. In fact, it got me banned from the API for data.gov.uk.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>