Write parser for the CKAN extras field
Instead of querying one dataset at a time I am using now the API
function current_package_list_with_resources for bulk import.

This, however, has to be done by using pagination. A limit and page
arguments specifies how much datasets every time should be fetched. Too
much datasets a time results in a timeout.

Another problem is that each field in the extras field is encoded as a
JSON meta-object. Which means the actual key and value are encoded as
properties. For that I had to write another parse method.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>