Add JSON-ZIP importer
Apparently, the newly introduced bulk importer with pagination does not
work with data.gov.uk. Often I receive internal server error responses.
There is, however, also a zip dump which can be used instead.

This is not only better, but also a lot faster. Due to the different
representations I also had to extract the name values from the group
and tags fields to create a new array.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>