Fix integer division bug
When computing the number of steps for querying the datasets I have
forgotten to cast one of the numbers to a float which resulted for
two repositories with few datasets in zero steps. Zero steps also
implied no dataset queries at all.

This was the problem, why I had no datasets in my database for certain
repositories.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>