Re-add catalog.data.gov to the repository list
I do not know what has changed, firstly I noticed that the CKAN instance
can be harvested without problems if I lower the limit size a lot, to 1
or 2. I thought this solved my problem, but now I notices that 50 or 100
works just as well.

I hope this was simply a temporary problem or they upgraded something.
For now I am leaving the limit at 100 as it seems quite efficient and
does not spam my log.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>