Add exception handling in case of blocked request
If there is a problem with the response, for instance I get forwarded to
a HTML page, then there will be a parsing error from the JSON module. I
have added exception handling which logs the occurrence and retries
after a couple of seconds. With every retry the sleep duration is
increased.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>