Continue refactoring of whole harvester structure
It is hard to decide at which parts changes have been made. In any case
I am still relying on Ruby libraries for performing the decompression.
This is still better, then doing that based on system commands. I have
replaced the zipruby library with the rubyzip library. The latter one
seems a lot more advanced. I hope, with this implementation, that this
is sufficiently memory efficient.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>