Update Sidekiq start-up handling
Instead of the cluttered Sidekiq start-up script I have outsourced the
configuration into its own YAML file. The Sidekiq process will now be
started as a daemon process from now on. For this I also need to have a
tmp directory to store the pid file and a log directory to store the log
file.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>