Add Yajl::Gzip stream reader and writer
For reasons unknown the Yajl stream writer and reader for compression
support are deprecated and will be removed with version 2.

There was no commit since 2011. Since I want to avoid the deprecation
message and because who knows if this ever becomes true, I have added
the wrapper to my own code.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>