Add disabled option to repositories list
With the US government shutdown I have to disable catalog.data.gov
again. To avoid the annoying commenting in and out pattern I have
added a disabled option.

With this commit I have also overhauled the main script. Now a list of
repositories can be passed to the repositories option.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>