Fix faulty pagination arguments
The arguments that were passed to perform the CKAN pagination were just
plain wrong. The starts parameter does not take a "page" argument, but
the actual offset in number of datasets.

Also the last query's rows argument is truncated to the remaining number
of datasets. This fixes the problem with the Swiss data portal. Instead
of 1630 datasets the harvester retrieved 2000 datasets including 370
duplicates.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>