Revise dump handler implementations and tests
I have learned that the value returned by the hash_start and array_start
callback methods is used for further references on the hash/array. This
means that I can drop the stack approach altogether.

This way the implementation is a lot easier. I have also revised the
next. Since this works very well now, the next step is to get the
callback handler working. Currently, it is only invoked once.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>