Add new dump list format with dump URLs and dates
In order to develop the snapshot analysis over time further I have
created a script which compiles a list of dump URLs from data.gov.uk.
In addition, it maps the URL to the date when the dump was created. In
order to import this I had to write new methods for the harvester. This
implementation is very experimental.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>